feat: :sparkles: adding in NativeSequence
rewrote tests, added in NativeSequence. Everything is breaking, but the API is much better. Need to readd in support for arrays, NativeArrays, and lists.

BREAKING CHANGE: this whole repo got foobar'd. All the tests have been rewritten, and the extension methods have been removed to focus on getting Blinq working with a single struct for now.