feat: :tada: Now managing memory conventionally
Found out a great way to manage memory with the NativeSequence. All methods that transform the sequence dispose the old inputs. You must call "Copy" to copy those sequence values over. All methods that convert the sequence to a type other than sequence do not call dispose on the sequence. You manually call Dispose on the sequence after that.